Rename PauliSum module
