+++ src/braket/quantum_information/__init__.py
@@ -17,4 +17,4 @@ sums of Pauli operators.
 """
 
 from braket.quantum_information.pauli_string import PauliString  # noqa: F401
-from braket.quantum_information.pauli_string_sum import PauliStringSum, PauliSum  # noqa: F401
+from braket.quantum_information.pauli_sum import PauliStringSum, PauliSum  # noqa: F401
